+++ src/catalog/types/FindOneProductQueryOutput.ts
@@ -1,0 +1,9 @@
+import { PriceBandType, ProductType } from './catalogGraphQLTypes';
+
+export type FindOneProductQueryOutput = {
+  product: ProductType;
+};
+
+export type FindOnePriceBandQueryOutput = {
+  priceBand: PriceBandType;
+};